requirements.txt
openai
langchain
faiss-cpu
pypdf
tiktoken
python-dotenv
pyodbc
streamlit
wget
azure-search-documents==11.4.0b8
azure-identity
PyPDF2
re